CB-3704. Validate file name (#2065)
* CB-3704. Validate file name

* CB-3704. Revert imports

* CB-3704. Refactor after review

* CB-3704. Fixed error message if files not created.

* CB-3074. Remove constant to Servlet

+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/server/events/WSDeleteTempFileHandler.java
@@ -25,6 +25,7 @@ import org.jkiss.dbeaver.model.websocket.event.WSEventDeleteTempFile;
 import org.jkiss.utils.IOUtils;
 
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 
 public class WSDeleteTempFileHandler implements WSEventHandler<WSEventDeleteTempFile> {
@@ -36,10 +37,12 @@ public class WSDeleteTempFileHandler implements WSEventHandler<WSEventDeleteTemp
         Path path = CBPlatform.getInstance()
                 .getTempFolder(new VoidProgressMonitor(), TEMP_FILE_FOLDER)
                 .resolve(sessionId);
-        try {
-            IOUtils.deleteDirectory(path);
-        } catch (IOException e) {
-            log.error("Error deleting temp path", e);
+        if (Files.exists(path)) {
+            try {
+                IOUtils.deleteDirectory(path);
+            } catch (IOException e) {
+                log.error("Error deleting temp path", e);
+            }
         }
     }
